@@ -17,18 +17,18 @@
     "build:dev-dry": "netlify build --dry"
   },
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-scripts": "5.0.0",
     "chart.js": "^3.7.1",
     "chartjs-plugin-datalabels": "^2.0.0",
     "msw": "^0.39.2",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
     "react-chartjs-2": "^4.0.1",
-    "react-dom": "^17.0.2",
     "react-router-dom": "^6.2.1",
-    "react-scripts": "4.0.3",
     "styled-components": "^5.3.3",
     "styled-icons": "^10.45.0",
     "suncalc": "^1.9.0",
